Fix compile for pono and deps for win32

+++ default/scripts/cvc4.sh
@@ -1,12 +1,17 @@
 cd cvc4
 export MACHINE_TYPE=x86_64
-if [ ${ARCH_BASE} != 'darwin' ]; then
+if [ ${ARCH_BASE} != 'darwin' ] &&  [ ${ARCH_BASE} != 'windows' ]; then
     sed -i -re 's,rm -rf src/antlr3debughandlers.c \&\& touch src/antlr3debughandlers.c,rm -rf src/antlr3debughandlers.c \&\& touch src/antlr3debughandlers.c \&\& cp  /usr/share/misc/config.* . ,g' ./contrib/get-antlr-3.4
 fi
 ANTLR_CONFIGURE_ARGS="--host=${CROSS_NAME} --build=`gcc -dumpmachine`"  ./contrib/get-antlr-3.4
 git clone https://github.com/uiri/toml.git
 export PYTHONPATH=$PYTHONPATH:`pwd`/toml
-sed -i -re 's,cmake \"\$root_dir\" \$cmake_opts,cmake \"\$root_dir\" \$cmake_opts -DCMAKE_TOOLCHAIN_FILE=\$\{CMAKE_TOOLCHAIN_FILE\},g' configure.sh
+if [ ${ARCH_BASE} == 'windows' ]; then
+    export CMAKE_TOOLCHAIN_FILE=${PATCHES_DIR}/Toolchain-mingw64.cmake
+    sed -i -re 's,cmake \"\$root_dir\" \$cmake_opts,cmake \"\$root_dir\" \$cmake_opts -DCMAKE_TOOLCHAIN_FILE=\$\{CMAKE_TOOLCHAIN_FILE\} -DCVC4_WINDOWS_BUILD=TRUE,g' configure.sh
+else
+    sed -i -re 's,cmake \"\$root_dir\" \$cmake_opts,cmake \"\$root_dir\" \$cmake_opts -DCMAKE_TOOLCHAIN_FILE=\$\{CMAKE_TOOLCHAIN_FILE\},g' configure.sh
+fi
 CXXFLAGS=-fPIC CFLAGS=-fPIC ./configure.sh --static --no-static-binary --prefix=${INSTALL_PREFIX}
 cd build
 make DESTDIR=${OUTPUT_DIR} -j${NPROC}
